서비스단 리팩토링 실습 정리

+++ src/main/java/com/sparta/java02/domain/category/dto/CategoryProductDTO.java
@@ -1,6 +1,7 @@
 package com.sparta.java02.domain.category.dto;
 
 import com.querydsl.core.annotations.QueryProjection;
+import java.math.BigDecimal;
 import lombok.Getter;
 
 @Getter
@@ -8,11 +9,12 @@ public class CategoryProductDTO {
 
   private final String categoryName;
   private final String productName;
-  private final Double price;
+  private final BigDecimal price;
   private final Integer stock;
 
   @QueryProjection //빌드 시 해당 객체를 Q 객체화
-  public CategoryProductDTO(String categoryName, String productName, Double price, Integer stock) {
+  public CategoryProductDTO(String categoryName, String productName, BigDecimal price,
+      Integer stock) {
     this.categoryName = categoryName;
     this.productName = productName;
     this.price = price;
